ci(release): throw if uncommitted changes in working tree

+++ scripts/release.ts
@@ -1,13 +1,21 @@
-import { spawnSync } from "node:child_process";
+import { spawnSync, SpawnSyncOptions } from "node:child_process";
 import fs from "node:fs";
 
 const args = process.argv.slice(2);
 
-fs.rmSync("dist", { recursive: true, force: true });
-
-function run(cmd: string, passedArgs: string[]): void {
-  spawnSync(cmd, passedArgs, { stdio: "inherit" });
+function run(
+  cmd: string,
+  passedArgs: string[],
+  options: SpawnSyncOptions = { stdio: "inherit" }
+) {
+  return spawnSync(cmd, passedArgs, options);
 }
+
+if (run("git", ["status", "-s"], { stdio: "pipe" }).stdout.toString()) {
+  throw new Error("Please commit all changes before running this script.");
+}
+
+fs.rmSync("dist", { recursive: true, force: true });
 
 // This is here purely to check the build works before triggering a release
 run("pnpm", ["build"]);
